+++ dune/gfe/cosseratvtkreader.hh
@@ -1,7 +1,9 @@
 #ifndef DUNE_GFE_COSSERATVTKREADER_HH
 #define DUNE_GFE_COSSERATVTKREADER_HH
 
-#include <dune/gfe/spaces/rigidbodymotion.hh>
+#include <dune/gfe/spaces/productmanifold.hh>
+#include <dune/gfe/spaces/realtuple.hh>
+#include <dune/gfe/spaces/rotation.hh>
 
 namespace Dune
 {
@@ -13,7 +15,7 @@ namespace Dune
     {
     public:
 
-      static void read(std::vector<RigidBodyMotion<double,3> >& configuration,
+      static void read(std::vector<ProductManifold<RealTuple<double,3>,Rotation<double,3> > >& configuration,
                        const std::string& filename)
       {
         VTKFile vtkFile;
@@ -23,14 +25,14 @@ namespace Dune
 
         for (size_t i=0; i<configuration.size(); i++)
         {
-          configuration[i].r = vtkFile.points_[i];
+          configuration[i][Indices::_0].globalCoordinates() = vtkFile.points_[i];
 
           FieldMatrix<double,3,3> R;
           for (int j=0; j<3; j++)
             for (int k=0; k<3; k++)
               R[j][k] = vtkFile.directors_[k][i][j];
 
-          configuration[i].q.set(R);
+          configuration[i][Indices::_1].set(R);
         }
 
       }